Handle optional sorting of errors before reporting
Users can specify now a command-line option to sort errors before they
are reported to stdout. When no option is given, errors are reported as
they are generated, preserving backwards compatibility. When given, a
comma-delimited list of field names must be given, with errors being
sorted by the first field, then the second, and so on.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>